LayoutTests/canvas/philip/tests/type.prototype-expected.txt

< [index] >
type.prototype
window.HTMLCanvasElement has prototype, which is { ReadOnly, DontDelete }. prototype has getContext, which is not
References: canvas.type
Defined in "Language Bindings for DOM Specifications" (draft)
Actual output:
Passed
